units to multiples
Changed the units module to return an array of 'multiples' objects that contain information that is needed for output purposes.
Since the sorting module was not terribly complex, I simply moved it over into the search module and modified when to sort by quantity then difference versus just the difference (when all results are outside the 10% margin for the latter).
Modified the index and output modules so that when running via index, a nice table is printed to the console. Perhaps being able to run npm start in the future could improve the look of the output.