possible working search algorithm
This push could use some polish, but I believe I have a working search algorithm that also combines multiples to provide additional combinations beyond simple multiples.
units.js - removed the algorithm for creating multiples and simply export units available.
search.js - implemented the multiples algorithm into this code to be able to consider the target and filter out multiples too high to be used. all multiples under 110% of the target are viable options and only those that also are above 90% of the target will be included as a result. all others are added as sub-optimal multiples to be combined with others later. all sub-optimal multiples will be considered to form combinations that result in a sum within the 10% margins.
output.js - modified the output formatting to be the same whether we used one unit value, multiples, or combinations to form the result.
index.js - minor modifications to match changes in the other files.